+++ src/index.js
@@ -1,4 +1,6 @@
 import SVGMap from './svg-map';
+import CheckboxSVGMap from './checkbox-svg-map';
+import RadioSVGMap from './radio-svg-map';
 import Australia from './maps/australia';
 import France from './maps/france';
 import Taiwan from './maps/taiwan';
@@ -8,6 +10,8 @@ import './svg-map.scss';
 
 export {
 	SVGMap,
+	CheckboxSVGMap,
+	RadioSVGMap,
 	Australia,
 	France,
 	Taiwan,